Update http_server.py
I like your use of os.path.split, which I hadn't considered using before. I think my code all comes to the same effect: basically we split the path, and then join it back together with a "webroot" at the beginning.

I removed the "if" in front of the FileNotFoundError. This works because if the path is either a directory or a file then we've already hit one of the return statements.